assignments refactoring
