+++ backend/main.py
@@ -2,6 +2,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 import routers.assignments
+import routers.submissions
 import routers.courses
 import routers.materials
 import routers.parents
@@ -12,6 +13,7 @@ import routers.users
 
 app = FastAPI()
 app.include_router(routers.assignments.router)
+app.include_router(routers.submissions.router)
 app.include_router(routers.courses.router)
 app.include_router(routers.materials.router)
 app.include_router(routers.parents.router)
